chore: move number b64 encoding to common
